refactor : 스케쥴 업데이트 관련 서비스 코드 분리 및 openAi 키 설정
